requirements.txt
fastapi
uvicorn
python-dotenv

openai

langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langchain-chroma

chromadb
